Break import cycles between Graph & graph_utils.py
