+++ fedot/core/dag/graph_utils.py
@@ -1,26 +1,11 @@
 from typing import Union, Sequence, List, TYPE_CHECKING, Callable
 
-from fedot.core.dag.convert import graph_structure_as_nx_graph
-from fedot.core.dag.graph_node import GraphNode
-
 if TYPE_CHECKING:
     from fedot.core.dag.graph import Graph
+    from fedot.core.dag.graph_node import GraphNode
 
 
-def get_nodes_degrees(graph: 'Graph') -> Sequence[int]:
-    """Nodes degree as the number of edges the node has:
-        ``degree = #input_edges + #out_edges``
-
-    Returns:
-        nodes degrees ordered according to the nx_graph representation of this graph
-    """
-    graph, _ = graph_structure_as_nx_graph(graph)
-    index_degree_pairs = graph.degree
-    node_degrees = [node_degree[1] for node_degree in index_degree_pairs]
-    return node_degrees
-
-
-def distance_to_root_level(graph: 'Graph', node: GraphNode) -> int:
+def distance_to_root_level(graph: 'Graph', node: 'GraphNode') -> int:
     """Gets distance to the final output node
 
     Args:
@@ -31,7 +16,7 @@ def distance_to_root_level(graph: 'Graph', node: GraphNode) -> int:
         int: distance to root level
     """
 
-    def recursive_child_height(parent_node: GraphNode) -> int:
+    def recursive_child_height(parent_node: 'GraphNode') -> int:
         node_child = graph.node_children(parent_node)
         if node_child:
             height = recursive_child_height(node_child[0]) + 1
@@ -42,11 +27,11 @@ def distance_to_root_level(graph: 'Graph', node: GraphNode) -> int:
     return height
 
 
-def distance_to_primary_level(node: GraphNode) -> int:
+def distance_to_primary_level(node: 'GraphNode') -> int:
     return node_depth(node) - 1
 
 
-def nodes_from_layer(graph: 'Graph', layer_number: int) -> Sequence[GraphNode]:
+def nodes_from_layer(graph: 'Graph', layer_number: int) -> Sequence['GraphNode']:
     """Gets all the nodes from the chosen layer up to the surface
 
     Args:
@@ -57,7 +42,7 @@ def nodes_from_layer(graph: 'Graph', layer_number: int) -> Sequence[GraphNode]:
         all nodes from the surface to the ``layer_number`` layer
     """
 
-    def get_nodes(node: Union[GraphNode, List[GraphNode]], current_height: int):
+    def get_nodes(node: Union['GraphNode', List['GraphNode']], current_height: int):
         """Gets all the parent nodes of ``node``
 
         :param node: node to get all subnodes from
@@ -101,7 +86,7 @@ def ordered_subnodes_hierarchy(node: 'GraphNode') -> List['GraphNode']:
     return subtree_impl(node)
 
 
-def node_depth(node: GraphNode) -> int:
+def node_depth(node: 'GraphNode') -> int:
     """Gets this graph depth from the provided ``node`` to the graph source node
 
     Args:
